# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Foundation documentation",
        "description": "Demo Todo List Api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8035/api",
            "description": "Laravel swagger API server"
        }
    ],
    "paths": {
...
        "/user/{user}/car-pick/{car}": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Pick car for user by ID",
                "operationId": "pickCar",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "The ID of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "car",
                        "in": "path",
                        "description": "The ID of Car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    }
                }
            }
        }
    },
...
}